+++ ios/Classes/core/PMAssetPathEntity.m
@@ -30,7 +30,7 @@
 @implementation PMAssetEntity {
 
 }
-- (instancetype)initWithId:(NSString *)id createDt:(long)createDt width:(int)width height:(int)height
+- (instancetype)initWithId:(NSString *)id createDt:(long)createDt width:(NSUInteger)width height:(NSUInteger)height
                   duration:(long)duration type:(int)type {
     self = [super init];
     if (self) {
@@ -45,7 +45,7 @@
     return self;
 }
 
-+ (instancetype)entityWithId:(NSString *)id createDt:(long)createDt width:(int)width height:(int)height
++ (instancetype)entityWithId:(NSString *)id createDt:(long)createDt width:(NSUInteger)width height:(NSUInteger)height
                     duration:(long)duration type:(int)type {
     return [[self alloc] initWithId:id createDt:createDt width:width height:height duration:duration type:type];
 }
